break up the main test file
